new greeter calling another service
