--- greeter.go
@@ -1,28 +1,0 @@
-package main
-
-import (
-	"fmt"
-	"net/http"
-	"os"
-)
-
-func main() {
-	fmt.Println("Hivemind's Go Greeter")
-	fmt.Println("You are running the service with this tag: ", os.Getenv("HELLO_TAG"))
-	http.HandleFunc("/", HelloServer)
-	http.ListenAndServe(":8080", nil)
-}
-
-func HelloServer(w http.ResponseWriter, r *http.Request) {
-	fmtStr := fmt.Sprintf("Hello, %s! I'm %s", GetIPFromRequest(r), os.Getenv("HOSTNAME"))
-	fmt.Println(fmtStr)
-	fmt.Fprintln(w, fmtStr)
-}
-
-func GetIPFromRequest(r *http.Request) string {
-	if fwd := r.Header.Get("x-forwarded-for"); fwd != "" {
-		return fwd
-	}
-
-	return r.RemoteAddr
-}
